#60 Moved config setting and applied setting to activity list
