+++ src/main/java/org/stt/config/ReportConfig.java
@@ -4,15 +4,6 @@ import java.time.Duration;
 
 public class ReportConfig implements ConfigurationContainer {
     private Duration roundDurationsTo = Duration.ofMinutes(5);
-    private boolean groupItems = true;
-
-    public boolean isGroupItems() {
-        return groupItems;
-    }
-
-    public void setGroupItems(boolean groupItems) {
-        this.groupItems = groupItems;
-    }
 
     public Duration getRoundDurationsTo() {
         return roundDurationsTo;
